Move ScanResultInterface implementation to RxBleScanResultMock and revert ScanResult changes
