+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/scan/ScanResult.java
@@ -3,10 +3,9 @@ package com.polidea.rxandroidble2.scan;
 import androidx.annotation.NonNull;
 
 import com.polidea.rxandroidble2.RxBleDevice;
-import com.polidea.rxandroidble2.internal.ScanResultInterface;
 import com.polidea.rxandroidble2.internal.logger.LoggerUtil;
 
-public class ScanResult implements ScanResultInterface {
+public class ScanResult {
 
     private final RxBleDevice bleDevice;
     private final int rssi;
@@ -34,22 +33,12 @@ public class ScanResult implements ScanResultInterface {
         return timestampNanos;
     }
 
-    @Override
-    public ScanCallbackType getScanCallbackType() {
+    public ScanCallbackType getCallbackType() {
         return callbackType;
     }
 
     public ScanRecord getScanRecord() {
         return scanRecord;
-    }
-
-    public String getAddress() {
-        return bleDevice == null ? null : bleDevice.getMacAddress();
-    }
-
-    public String getDeviceName() {
-        RxBleDevice device = getBleDevice();
-        return device == null ? null : device.getName();
     }
 
     @Override
